Beginning of Tucson (Working Title)
This is a song I came up with in a dream last night. I was in a car driving past a low rise building with a Jim Beam poster on the side. I thought of Tucson Arizona and woke up with this song in my head and the idea to start an open source project to write songs. Still trying to figure out how to add recordings associated with each commit, that is going to be important. I'll figure that out.